move tools closer to workflow for kol
