+++ auto-agents-framework/src/agents/workflows/kol/tools.ts
@@ -1,5 +1,5 @@
-import { TwitterAPI } from '../../services/twitter/client.js';
-import { createFetchTimelineTool } from '../tools/fetchTimelineTool.js';
+import { TwitterAPI } from '../../../services/twitter/client.js';
+import { createFetchTimelineTool } from '../../tools/fetchTimelineTool.js';
 
 export const createTools = (twitterAPI: TwitterAPI) => {
   const fetchTimelineTool = createFetchTimelineTool(twitterAPI);
